Struct tentacle_multiaddr::Multiaddr

pub struct Multiaddr { /* fields omitted */ }

Representation of a Multiaddr.

Methods

impl Multiaddr

pub fn is_empty(&self) -> bool

Returns true if the multiaddres has no data.

pub fn len(&self) -> usize

Return the length in bytes of this multiaddress.

pub fn to_vec(&self) -> Vec<u8>

Return a copy of this Multiaddr's byte representation.

pub fn push(&mut self, p: Protocol)

Adds an already-parsed address component to the end of this multiaddr.

Examples

use tentacle_multiaddr::{Multiaddr, Protocol};

let mut address: Multiaddr = "/ip4/127.0.0.1".parse().unwrap();
address.push(Protocol::TCP(10000));
println!("{}", address);
assert_eq!(address, "/ip4/127.0.0.1/tcp/10000".parse().unwrap());

pub fn pop<'a>(&mut self) -> Option<Protocol<'a>>

Pops the last Protocol of this multiaddr, or None if the multiaddr is empty.

use std::net::Ipv4Addr;
use tentacle_multiaddr::{Multiaddr, Protocol};

let mut address: Multiaddr = "/ip4/127.0.0.1/tcp/5678".parse().unwrap();

assert_eq!(address.pop().unwrap(), Protocol::TCP(5678));
assert_eq!(address.pop().unwrap(), Protocol::IP4(Ipv4Addr::new(127, 0, 0, 1)));

pub fn iter(&self) -> Iter

Returns the components of this multiaddress.

use std::net::Ipv4Addr;
use tentacle_multiaddr::{Multiaddr, Protocol};

let address: Multiaddr = "/ip4/127.0.0.1/tcp/5678".parse().unwrap();

let components = address.iter().collect::<Vec<_>>();
assert_eq!(components[0], Protocol::IP4(Ipv4Addr::new(127, 0, 0, 1)));
assert_eq!(components[1], Protocol::TCP(5678));

pub fn replace<'a, F>(&self, at: usize, by: F) -> Option<Multiaddr> where
F: FnOnce(&Protocol) -> Option<Protocol<'a>>,

Replace a Protocol at some position in this Multiaddr.

The parameter at denotes the index of the protocol at which the function by will be applied to the current protocol, returning an optional replacement.

If at is out of bounds or by does not yield a replacement value, None will be returned. Otherwise a copy of this Multiaddr with the updated Protocol at position at will be returned.